Win32 doesn't know "roundf()" - reverted to "round()" - caught by BCFH|Ärkefiende
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk/ac@5242 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ source/src/rendergl.cpp
@@ -1027,7 +1027,7 @@ void gl_drawframe(int w, int h, float changelod, float curfps)
     }
 
     extern vector<vertex> verts;
-    gl_drawhud(w, h, (int)roundf(curfps), nquads, verts.length(), underwater);
+    gl_drawhud(w, h, (int)round(curfps), nquads, verts.length(), underwater);
 
     glEnable(GL_CULL_FACE);
     glEnable(GL_FOG);
